Move screensaver impls to new module and disable take_next
