+++ src/api/canon.rs
@@ -7,9 +7,7 @@ use crate::{
         image_dimensions::{self, FetchImageDimensionsError},
         IMAGES_DIR,
     },
-    domain::{
-        actions::image::UpdateCanon, models::Image, screen_saver_manager::ScreenSaverManager,
-    },
+    domain::{actions::image::UpdateCanon, models::Image, screensaver::Screensaver},
 };
 
 #[derive(Debug, Serialize)]
@@ -104,7 +102,7 @@ impl From<FetchCanonError> for UpdateCanonError {
 
 pub async fn update_canon(
     update_canon_op: &impl UpdateCanon,
-    mngr: &mut ScreenSaverManager,
+    mngr: &mut impl Screensaver,
 ) -> Result<(), UpdateCanonError> {
     let images = fetch_canon()?;
     update_canon_op
