+++ src/domain.rs
@@ -1,3 +1,3 @@
 pub mod actions;
 pub mod models;
-pub mod screen_saver_manager;
+pub mod screensaver;
