+++ src/state.rs
@@ -1,0 +1,2 @@
+pub mod screensaver_manager;
+mod screensaver_state;
